Cifar10-Image-Classification-and-Deployment/requirements.txt
flask
torch
torchvision
Pillow
numpy
tqdm

